requirements.txt
black==22.3.*
environs==9.5.*
flake8==4.0.*
python-telegram-bot==13.12
redis==4.3.*
